+++ tw-context-ownership-starter/src/main/java/com/transferwise/common/context/ownership/EntryPointOwnerProviderRegistry.java
@@ -1,0 +1,5 @@
+package com.transferwise.common.context.ownership;
+
+public interface EntryPointOwnerProviderRegistry {
+  String getOwner(String entryPointGroup, String entryPointName);
+}