+++ tw-context-ownership-starter/src/test/java/com/transferwise/common/context/ownership/TestApplication.java
@@ -1,0 +1,8 @@
+package com.transferwise.common.context.ownership;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TestApplication {
+
+}