+++ tw-context/src/main/java/com/transferwise/common/context/TwContext.java
@@ -1,32 +1,40 @@
 package com.transferwise.common.context;
 
+import com.google.common.util.concurrent.RateLimiter;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.concurrent.CopyOnWriteArrayList;
+import java.util.Set;
+import java.util.concurrent.CopyOnWriteArraySet;
 import java.util.function.Function;
 import java.util.function.Supplier;
 import lombok.Getter;
 import lombok.NonNull;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.MDC;
 
+@Slf4j
 public class TwContext {
 
   public static final String GROUP_KEY = "TwContextGroup";
   public static final String NAME_KEY = "TwContextName";
+  public static final String OWNER_KEY = "TwContextOwner";
   public static final String GROUP_GENERIC = "Generic";
   public static final String NAME_GENERIC = "Generic";
+  public static final String OWNER_GENERIC = "Generic";
 
   public static final String MDC_KEY_EP_NAME = "tw_entrypoint_name";
   public static final String MDC_KEY_EP_GROUP = "tw_entrypoint_group";
+  public static final String MDC_KEY_EP_OWNER = "tw_entrypoint_owner";
 
   private static final ThreadLocal<TwContext> contextTl = new ThreadLocal<>();
-  private static final List<TwContextExecutionInterceptor> interceptors =
-      new CopyOnWriteArrayList<>();
+  private static final Set<TwContextExecutionInterceptor> interceptors = new CopyOnWriteArraySet<>();
+  private static final Set<TwContextAttributeChangeListener> attributeChangeListeners = new CopyOnWriteArraySet<>();
   private static final TwContext ROOT_CONTEXT = new TwContext(null, true);
+  private static final RateLimiter throwableLoggingRateLimiter = RateLimiter.create(2);
 
   public static TwContext current() {
     TwContext twContext = contextTl.get();
@@ -37,8 +45,24 @@ public class TwContext {
     interceptors.add(interceptor);
   }
 
-  public static List<TwContextExecutionInterceptor> getExecutionInterceptors() {
+  public static boolean removeExecutionInterceptor(@NonNull TwContextExecutionInterceptor interceptor) {
+    return interceptors.remove(interceptor);
+  }
+
+  public static Set<TwContextExecutionInterceptor> getExecutionInterceptors() {
     return interceptors;
+  }
+
+  public static void addAttributeChangeListener(@NonNull TwContextAttributeChangeListener listener) {
+    attributeChangeListeners.add(listener);
+  }
+
+  public static boolean removeAttributeChangeListener(TwContextAttributeChangeListener listener) {
+    return attributeChangeListeners.remove(listener);
+  }
+
+  public static Set<TwContextAttributeChangeListener> getAttributeChangeListeners() {
+    return attributeChangeListeners;
   }
 
   public static void putCurrentMdc(@NonNull String key, String value) {
@@ -153,29 +177,40 @@ public class TwContext {
   }
 
   @SuppressWarnings("UnusedReturnValue")
-  public TwContext put(String key, Object value) {
+  public TwContext put(String key, Object newValue) {
     if (root) {
       throw new IllegalStateException("You can not add values into root context.");
     }
-    if (value == null) {
-      attributes.remove(key);
-      newAttributes.remove(key);
-    } else {
-      if (!Objects.equals(attributes.get(key), value)) {
-        attributes.put(key, value);
-        newAttributes.put(key, value);
+    Object oldValue = attributes.get(key);
+
+    if (!Objects.equals(oldValue, newValue)) {
+      if (newValue == null) {
+        attributes.remove(key);
+        newAttributes.remove(key);
+      } else {
+        attributes.put(key, newValue);
+        newAttributes.put(key, newValue);
       }
+      fireAttributeChangeEvent(key, oldValue, newValue);
     }
+
     return this;
   }
 
   @SuppressWarnings("UnusedReturnValue")
   public TwContext setName(@NonNull String group, @NonNull String name) {
-    put(NAME_KEY, name);
     put(GROUP_KEY, group);
+    put(NAME_KEY, name);
 
     putMdc(MDC_KEY_EP_GROUP, group);
     putMdc(MDC_KEY_EP_NAME, name);
+
+    return this;
+  }
+
+  public TwContext setOwner(@NonNull String owner) {
+    put(OWNER_KEY, owner);
+    putMdc(MDC_KEY_EP_OWNER, owner);
 
     return this;
   }
@@ -188,6 +223,11 @@ public class TwContext {
   public String getGroup() {
     String group = get(GROUP_KEY);
     return group == null ? GROUP_GENERIC : group;
+  }
+
+  public String getOwner() {
+    String owner = get(OWNER_KEY);
+    return owner == null ? OWNER_GENERIC : owner;
   }
 
   @SuppressWarnings("unchecked")
@@ -247,5 +287,19 @@ public class TwContext {
     }
     return interceptors.get(interceptorIdx)
         .intercept(this, () -> executeWithInterceptors(supplier, interceptors, interceptorIdx + 1));
+  }
+
+  private void fireAttributeChangeEvent(String key, Object oldValue, Object newValue) {
+    if (attributeChangeListeners != null) {
+      try {
+        attributeChangeListeners.forEach((l) -> l.attributeChanged(this, key, oldValue, newValue));
+      } catch (Throwable t) {
+        // This is just a safety net, every listener needs to be bullet-proof by themselves.
+        if (throwableLoggingRateLimiter.tryAcquire()) {
+          // Don't log value, could be PII.
+          log.error("Attribute change listener failed for key '" + key + "'.", t);
+        }
+      }
+    }
   }
 }