+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2020-05-20
+### Added
+- TwContext now has entry point "Owner" as first class citizen, together with name and group.  
+
+- tw-context-ownership-starter module.
+A library to automatically set entry point owner, based on configuration or "handler" classes.
+
 ## [0.2.2] - 2020-05-20
 ### Changed
 - Deadline exception also contains information how long has passed from the start of the unit of work.